+++ inst/include/colourvalues/alpha/alpha.hpp
@@ -22,10 +22,13 @@ namespace alpha {
     } else if ( alpha_size == x_size ) {
       return ALPHA_VECTOR;
     }
-    return ALPHA_UNKNOWN;
+    return ALPHA_UNKNOWN;    // #nocov
   }
 
-  inline Rcpp::NumericVector validate_alpha( Rcpp::NumericVector& alpha, int& alpha_type, int& x_size ) {
+  inline Rcpp::NumericVector validate_alpha( Rcpp::NumericVector& a, int& alpha_type, int& x_size ) {
+
+    // Issue 47
+    Rcpp::NumericVector alpha = Rcpp::clone( a );
 
     if ( alpha_type == ALPHA_CONSTANT ) {
       if ( alpha[0] >= 0 && alpha[0] < 1 ) {
@@ -61,10 +64,10 @@ namespace alpha {
       }
       return alpha;
 
-    } else if ( alpha_type == ALPHA_UNKNOWN ) {
-      Rcpp::stop("Unknown alpha definition");
+    } else if ( alpha_type == ALPHA_UNKNOWN ) {                // #nocov
+      Rcpp::stop("colourvalues - Unknown alpha definition");   // #nocov
     }
-    return 0;  // never reached
+    return 0;  // #nocov never reached
   }
 
 } // namespace alpha